Add print_tree() introspection method
print_tree() writes a textual representation of the object and all of its
children (optionally with a depth limit) to stdout, a file object, or a file
name.

This is easier to use than "autopilot vis" when there are deeply nested
structures, or in cases where vis is not available.