bamf: Don't inherit our stdout to launched applications
We want to be able to tee autopilot's output to a file, so we must not inherit
our stdout to subprocesses. Gio.DesktopAppInfo.launch_uris() does that by
default, so call launch_uris_as_manager() instead for more fine-grained
control, which allows us to redirect stdout.

Bump gir1.2-glib-2.0 dependency to ensure proper GI annotations on that method.